Add alt way to copy text. Add download button on Windows.
Add another way to copy passphrase to clipboard if navigator.clipboard is
not supported in the browser by adding a new textarea, passing that the
text, then copying that textarea before deleting it, it should not be seen
by the user.
Display download button to Windows users. This links to the latest release
page on the github repository.